+++ srcs/iambot.cpp
@@ -6,14 +6,14 @@
 /*   By: caguillo <caguillo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/18 00:21:15 by caguillo          #+#    #+#             */
-/*   Updated: 2025/04/24 15:11:03 by caguillo         ###   ########.fr       */
+/*   Updated: 2025/04/24 23:22:22 by caguillo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 # include "Server.hpp"
 
 // IAMBOT keybot
-void Server::iambot(std::vector<std::string>& tab_msg, int clt_idx, int tab_idx)
+void Server::iambot(std::vector<std::string>& tab_msg, int clt_idx, size_t tab_idx)
 {       
     size_t i = tab_idx + 1;
