+++ srcs/topic.cpp
@@ -6,14 +6,14 @@
 /*   By: caguillo <caguillo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/08 13:50:52 by caguillo          #+#    #+#             */
-/*   Updated: 2025/04/24 15:05:00 by caguillo         ###   ########.fr       */
+/*   Updated: 2025/04/24 23:24:03 by caguillo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "Server.hpp"
 
 // Topic #test : --> clear the topic
-void Server::topic(std::vector<std::string>& tab_msg, int clt_idx, int tab_idx)
+void Server::topic(std::vector<std::string>& tab_msg, int clt_idx, size_t tab_idx)
 {
     size_t i = tab_idx + 1;
